changed structure and implemented service. Now we can go on and change add button
