--- src/app/components/footprint-items/footprint-items.component.html
@@ -1,1 +1,0 @@
-<p>footprint-items works!</p>
