+++ src/app/services/itemservice.service.ts
@@ -18,4 +18,39 @@ export class ItemserviceService {
   changeItem(item: Item) {
     this.messageSource.next(item)
   }
+
+  getItems(){
+    return [
+      {
+        id: 1,
+        name: 'Avocado',
+        co2value: 100,
+      },
+      {
+        id: 2,
+        name: 'Banane',
+        co2value: 70,
+      },
+      {
+        id: 3,
+        name: 'Autofahrt',
+        co2value: 200,
+      },
+      {
+        id: 4,
+        name: 'Heizung',
+        co2value: 20,
+      },
+      {
+        id: 5,
+        name: 'GasHerd',
+        co2value: 50,
+      },
+      {
+        id: 6,
+        name: 'Tomaten',
+        co2value: 20,
+      },
+    ]
+  }
 }
